refactor : PathFinder validation안에서 getPath 하는부분 제거
